refactor workbench analytics route
