+++ src/app/router_registry.py
@@ -223,6 +223,7 @@ from app.routers.reporting_schedules import schedules_router as reporting_schedu
 from app.routers.reporting_snapshots import router as reporting_snapshots_router
 from app.routers.source_products import router as source_products_router
 from app.routers.workbench import router as workbench_router
+from app.routers.workbench_analytics import router as workbench_analytics_router
 from app.routers.workbench_performance import router as workbench_performance_router
 from app.routers.workbench_performance_advisor_brief import (
     router as workbench_performance_advisor_brief_router,
@@ -360,6 +361,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_wave_campaign_lifecycle_router)
     app.include_router(dpm_wave_campaign_readiness_router)
     app.include_router(workbench_router)
+    app.include_router(workbench_analytics_router)
     app.include_router(workbench_performance_router)
     app.include_router(workbench_performance_details_router)
     app.include_router(workbench_performance_advisor_brief_router)
